Hide private Mock methods.
Converted the private Mock methods (those that start with _) to
internal functions not tied to the Mock object.  This keeps Mock
uncluttered but does open the functions up for use by any other
code within the mock.coffee source file, which is perhaps not so
bad if we stick to one object per file or have a particularly
simple application/library, as this one is.